Fixed wrong use of variable
File::Find does a chdir() when traversing a directory structure. The variable $File::Find::name that was passed to the print_todos function contains an entire relative path + filename which is "wrong" after File::Find has changed directory.

The $_ variable is now passed instead, which contains the file name only, so it can be found in the directory changed into.